renamed proxyHelper to propertyAssent
