+++ src/utilities/propertyAssent.mjs
